requirements.txt
h5py==3.8.0
numpy==1.24.2
pandas==2.0.1
torch==1.13.1
torchgeo==0.4.1
trimesh==3.22.0
python==3.9.6
```